Add files
Added R script for two functions:

msci.download() which accepts a number of parameters, including which
indices you wish to download, start/end dates, etc.

msci.list() which gives back a table of all the download codes that are
included. Note: the result of this function is the same as the content
in the "MSCI Download Codes.csv" file.